Revert "SNOW-1163203: Increased Max LOB size in metadata (#1806)"
This reverts commit d4504f8b7780a2504712b21569bb29ccaf755073.
