+++ src/test/java/net/snowflake/client/jdbc/DatabaseMetaDataLatestIT.java
@@ -3,8 +3,6 @@
  */
 package net.snowflake.client.jdbc;
 
-import static net.snowflake.client.jdbc.DatabaseMetaDataIT.EXPECTED_MAX_BINARY_LENGTH;
-import static net.snowflake.client.jdbc.DatabaseMetaDataIT.EXPECTED_MAX_CHAR_LENGTH;
 import static net.snowflake.client.jdbc.DatabaseMetaDataIT.verifyResultSetMetaDataColumns;
 import static net.snowflake.client.jdbc.SnowflakeDatabaseMetaData.NumericFunctionsSupported;
 import static net.snowflake.client.jdbc.SnowflakeDatabaseMetaData.StringFunctionsSupported;
@@ -774,8 +772,8 @@ public class DatabaseMetaDataLatestIT extends BaseJDBCTest {
           "create or replace table JDBC_TBL111(colA string, colB decimal, colC " + "timestamp)");
       /* Create a UDF that returns a table made up of 4 columns from 2 different tables, joined together */
       statement.execute(
-          "create or replace function FUNC112 () RETURNS TABLE(colA string(16777216), colB decimal, bin2 "
-              + "binary(8388608) , sharedCol decimal) COMMENT= 'returns table of 4 columns' as 'select"
+          "create or replace function FUNC112 () RETURNS TABLE(colA string, colB decimal, bin2"
+              + " binary, sharedCol decimal) COMMENT= 'returns table of 4 columns' as 'select"
               + " JDBC_TBL111.colA, JDBC_TBL111.colB, BIN_TABLE.bin2, BIN_TABLE.sharedCol from"
               + " JDBC_TBL111 inner join BIN_TABLE on JDBC_TBL111.colB =BIN_TABLE.sharedCol'");
       DatabaseMetaData metaData = connection.getMetaData();
@@ -879,7 +877,7 @@ public class DatabaseMetaDataLatestIT extends BaseJDBCTest {
         assertEquals(DatabaseMetaData.functionNullableUnknown, resultSet.getInt("NULLABLE"));
         assertEquals("returns table of 4 columns", resultSet.getString("REMARKS"));
         // char octet length column is not supported and always returns 0
-        assertEquals(EXPECTED_MAX_CHAR_LENGTH, resultSet.getInt("CHAR_OCTET_LENGTH"));
+        assertEquals(16777216, resultSet.getInt("CHAR_OCTET_LENGTH"));
         assertEquals(1, resultSet.getInt("ORDINAL_POSITION"));
         // is_nullable column is not supported and always returns empty string
         assertEquals("", resultSet.getString("IS_NULLABLE"));
@@ -929,7 +927,7 @@ public class DatabaseMetaDataLatestIT extends BaseJDBCTest {
         assertEquals(DatabaseMetaData.functionNullableUnknown, resultSet.getInt("NULLABLE"));
         assertEquals("returns table of 4 columns", resultSet.getString("REMARKS"));
         // char octet length column is not supported and always returns 0
-        assertEquals(EXPECTED_MAX_BINARY_LENGTH, resultSet.getInt("CHAR_OCTET_LENGTH"));
+        assertEquals(8388608, resultSet.getInt("CHAR_OCTET_LENGTH"));
         assertEquals(3, resultSet.getInt("ORDINAL_POSITION"));
         // is_nullable column is not supported and always returns empty string
         assertEquals("", resultSet.getString("IS_NULLABLE"));
@@ -1224,8 +1222,8 @@ public class DatabaseMetaDataLatestIT extends BaseJDBCTest {
         statement.execute(
             "create or replace table "
                 + targetTable
-                + "(C1 int, C2 varchar(100), C3 string(16777216) default '', C4 number(18,4), C5 double,"
-                + " C6 boolean, C7 date not null, C8 time, C9 timestamp_ntz(7), C10 binary(8388608),C11"
+                + "(C1 int, C2 varchar(100), C3 string default '', C4 number(18,4), C5 double,"
+                + " C6 boolean, C7 date not null, C8 time, C9 timestamp_ntz(7), C10 binary,C11"
                 + " variant, C12 timestamp_ltz(8), C13 timestamp_tz(3))");
 
         DatabaseMetaData metaData = connection.getMetaData();
@@ -1292,14 +1290,14 @@ public class DatabaseMetaDataLatestIT extends BaseJDBCTest {
           assertEquals("C3", resultSet.getString("COLUMN_NAME"));
           assertEquals(Types.VARCHAR, resultSet.getInt("DATA_TYPE"));
           assertEquals("VARCHAR", resultSet.getString("TYPE_NAME"));
-          assertEquals(EXPECTED_MAX_CHAR_LENGTH, resultSet.getInt("COLUMN_SIZE"));
+          assertEquals(16777216, resultSet.getInt("COLUMN_SIZE"));
           assertEquals(0, resultSet.getInt("DECIMAL_DIGITS"));
           assertEquals(0, resultSet.getInt("NUM_PREC_RADIX"));
           assertEquals(ResultSetMetaData.columnNullable, resultSet.getInt("NULLABLE"));
           assertEquals("", resultSet.getString("REMARKS"));
           assertEquals("", resultSet.getString("COLUMN_DEF"));
 
-          assertEquals(EXPECTED_MAX_CHAR_LENGTH, resultSet.getInt("CHAR_OCTET_LENGTH"));
+          assertEquals(16777216, resultSet.getInt("CHAR_OCTET_LENGTH"));
           assertEquals(3, resultSet.getInt("ORDINAL_POSITION"));
           assertEquals("YES", resultSet.getString("IS_NULLABLE"));
           assertNull(resultSet.getString("SCOPE_CATALOG"));
@@ -1467,7 +1465,7 @@ public class DatabaseMetaDataLatestIT extends BaseJDBCTest {
           assertEquals("C10", resultSet.getString("COLUMN_NAME"));
           assertEquals(Types.BINARY, resultSet.getInt("DATA_TYPE"));
           assertEquals("BINARY", resultSet.getString("TYPE_NAME"));
-          assertEquals(EXPECTED_MAX_BINARY_LENGTH, resultSet.getInt("COLUMN_SIZE"));
+          assertEquals(8388608, resultSet.getInt("COLUMN_SIZE"));
           assertEquals(0, resultSet.getInt("DECIMAL_DIGITS"));
           assertEquals(0, resultSet.getInt("NUM_PREC_RADIX"));
           assertEquals(ResultSetMetaData.columnNullable, resultSet.getInt("NULLABLE"));
